+++ smart-contracts/contracts-common/concordium-contracts-common/src/types.rs
@@ -1048,6 +1048,21 @@ impl AccountAddress {
             None
         }
     }
+
+    /// Get the `n-th` alias of an address. There are 2^24 possible aliases.
+    /// If the counter is `>= 2^24` then this function will have unintended
+    /// behaviour, since it will wrap around. Meaning that counter values
+    /// 2^24 and 0 will give the same alias.
+    pub const fn get_alias_unchecked(&self, counter: u32) -> Self {
+        let mut data = self.0;
+        let counter_bytes = counter.to_be_bytes();
+
+        data[29] = counter_bytes[1];
+        data[30] = counter_bytes[2];
+        data[31] = counter_bytes[3];
+
+        Self(data)
+    }
 }
 
 /// Address of a contract.